refactor: centralizes navigation components
