+++ src/_components/navigation/thatConference/EventNav.svelte
@@ -1,37 +1,50 @@
 <script>
-	import { Shell as ShellButton } from '$elements/buttons';
-	import NavLink from './components/navLink.svelte';
+	export let event;
+
+	import { Shell as ShellButton, StandardShell } from '$elements/buttons';
+	import NavLink from '../_components/NavLink.svelte';
+
+	function buildUrl(page) {
+		return `/${event.slug}/${page}/`;
+	}
+
+	function buildAnchorLink(anchor) {
+		return `/${event.slug}#${anchor}`;
+	}
 </script>
 
-<div class="relative bg-white py-6">
-	<div class="mx-auto flex max-w-screen-xl items-center">
-		<div>
-			<img class="w-36" src="/images/ThatConferenceLogo.svg" alt="THAT Logo" />
+<div class=" bg-white py-12">
+	<div class="mx-auto flex max-w-screen-xl items-center justify-between">
+		<!-- logo -->
+		<div class="flex items-center">
+			<div class="cursor">
+				<a data-sveltekit-prefetch href="/{event.slug}/">
+					<img class="h-16" src="/images/Trees-Solid.svg" alt="THAT Logo" />
+				</a>
+			</div>
+
+			<!-- links -->
+			<div class="ml-8 flex space-x-4 text-lg font-bold">
+				<NavLink href={buildUrl('speakers')}>Speakers</NavLink>
+				<NavLink href={buildUrl('schedule')}>Schedule</NavLink>
+				<NavLink href={buildUrl('sponsors')}>Sponsors</NavLink>
+				<NavLink href={buildAnchorLink('dates')}>Dates</NavLink>
+				<NavLink href={buildUrl('travel')}>Travel</NavLink>
+				<NavLink href={buildAnchorLink('fieldGuide')}>Field Guide</NavLink>
+			</div>
 		</div>
 
-		<!-- links -->
-		<div class="ml-12 flex text-sm font-bold">
-			<NavLink href="/wi">Wisconsin</NavLink>
-			<NavLink href="/tx">Texas</NavLink>
-			<NavLink href="/partners">Sponsors</NavLink>
-			<NavLink href="/support/travel">Travel</NavLink>
-		</div>
-
-		<div class="flex w-full flex-1 justify-end space-x-4">
-			<div>
-				<ShellButton>
-					<div class="px-4 py-2">
-						<a href="/support/contact-us"> Contact Us </a>
-					</div>
-				</ShellButton>
-			</div>
-			<div>
-				<ShellButton>
-					<div class="px-4 py-2">
-						<a href="/sponsorships/apply/#prospectus"> Sponsorship Prospectus </a>
-					</div>
-				</ShellButton>
-			</div>
+		<div class="flex space-x-4">
+			<ShellButton>
+				<div class="px-4 py-2">
+					<a href="/sponsorships/apply/#prospectus">Sponsor Today</a>
+				</div>
+			</ShellButton>
+			<StandardShell>
+				<div class="px-4 py-2">
+					<a href={buildUrl('tickets')}>Purchase Tickets</a>
+				</div>
+			</StandardShell>
 		</div>
 	</div>
 </div>
