+++ src/_components/navigation/thatConference/MobileNavHat.svelte
@@ -1,5 +1,9 @@
 <script>
-	import LeftArrow from '$elements/svgs/LeftArrow.svelte';
+	export let event;
+
+	import { LeftArrow } from '$elements/svgs';
+
+	const [venue] = event.venues;
 </script>
 
 <div class="relative col-span-1 grid h-14 place-self-center bg-that-navy text-white">
@@ -7,7 +11,15 @@
 		<div class="font-semibold antialiased">
 			<a href="/" class="flex items-center space-x-2">
 				<LeftArrow />
-				<span>Back to thatconference.com</span>
+				<span>Back to THAT</span>
+			</a>
+		</div>
+
+		<div>
+			<a href={`/${event.slug}`}>
+				<div class="flex space-x-4 font-semibold uppercase tracking-wide antialiased">
+					<div class="uppercase">{`${venue.city}, ${venue.state}`}</div>
+				</div>
 			</a>
 		</div>
 	</div>
